+++ test/api_stub/requests/traffic_manager/traffic_manager_endpoint.rb
@@ -1,9 +1,9 @@
 module ApiStub
   module Requests
-    module Network
+    module TrafficManager
       # Mock class for raffic Manager End Point Requests
       class TrafficManagerEndPoint
-        def self.create_traffic_manager_endpoint_response
+        def self.create_traffic_manager_endpoint_response(traffic_manager_client)
           body = '{
             "id": "/subscriptions/######/resourceGroups/fog-test-rg/providers/Microsoft.Network/trafficManagerProfiles/fog-test-profile/externalEndpoints/fog-test-end-point?api-version=2015-11-01",
             "name": "fog-test-end-point",
@@ -17,7 +17,23 @@ module ApiStub
               "endpointLocation": "northeurope"
             }
           }'
-          body
+          endpoint_mapper = Azure::ARM::TrafficManager::Models::Endpoint.mapper
+          traffic_manager_client.deserialize(endpoint_mapper, JSON.load(body), 'result.body')
+        end
+
+        def self.endpoint_hash
+          {
+            resource_group: 'resource-group',
+            name: 'name',
+            traffic_manager_profile_name: 'traffic_manager_profile_name',
+            type: 'type',
+            target_resource_id: 'target_resource_id',
+            target: 'target',
+            weight: 'weight',
+            priority: 'priority',
+            endpoint_location: 'endpoint_location',
+            min_child_endpoints: 'min_child_endpoints'
+          }
         end
       end
     end